requirements.txt
attrs==23.2.0
hikari==2.0.0.dev126
hikari-lightbulb==2.3.5
hikari-miru==4.1.1
